+++ deploy/crds/apps.open-cluster-management.io_helmreleases_crd.yaml
@@ -1,9 +1,9 @@
 apiVersion: apiextensions.k8s.io/v1beta1
 kind: CustomResourceDefinition
 metadata:
-  name: helmreleases.multicloud-apps.io
+  name: helmreleases.apps.open-cluster-management.io
 spec:
-  group: multicloud-apps.io
+  group: apps.open-cluster-management.io
   names:
     kind: HelmRelease
     listKind: HelmReleaseList
