+++ pkg/apis/apps/group.go
@@ -14,9 +14,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package app contains app API versions.
+// Package apps contains app API versions.
 //
 // This file ensures Go source parsers acknowledge the app package
 // and any child packages. It can be removed if any other Go source files are
 // added to this package.
-package app
+package apps
