+++ pkg/apis/apps/v1/doc.go
@@ -16,5 +16,5 @@ limitations under the License.
 
 // Package v1 contains API Schema definitions for the app v1 API group
 // +k8s:deepcopy-gen=package,register
-// +groupName=multicloud-apps.io
+// +groupName=apps.open-cluster-management.io
 package v1
