+++ pkg/apis/apps/v1/zz_generated.openapi.go
@@ -25,12 +25,12 @@ import (
 
 func GetOpenAPIDefinitions(ref common.ReferenceCallback) map[string]common.OpenAPIDefinition {
 	return map[string]common.OpenAPIDefinition{
-		"./pkg/apis/multicloudapps/v1.HelmRelease":     schema_pkg_apis_app_v1_HelmRelease(ref),
-		"./pkg/apis/multicloudapps/v1.HelmReleaseRepo": schema_pkg_apis_app_v1_HelmReleaseRepo(ref),
+		"github.com/open-cluster-management/multicloud-operators-subscription-release/pkg/apis/apps/v1.HelmRelease":     schema_pkg_apis_apps_v1_HelmRelease(ref),
+		"github.com/open-cluster-management/multicloud-operators-subscription-release/pkg/apis/apps/v1.HelmReleaseRepo": schema_pkg_apis_apps_v1_HelmReleaseRepo(ref),
 	}
 }
 
-func schema_pkg_apis_app_v1_HelmRelease(ref common.ReferenceCallback) common.OpenAPIDefinition {
+func schema_pkg_apis_apps_v1_HelmRelease(ref common.ReferenceCallback) common.OpenAPIDefinition {
 	return common.OpenAPIDefinition{
 		Schema: spec.Schema{
 			SchemaProps: spec.SchemaProps{
@@ -58,28 +58,28 @@ func schema_pkg_apis_app_v1_HelmRelease(ref common.ReferenceCallback) common.Ope
 					},
 					"repo": {
 						SchemaProps: spec.SchemaProps{
-							Ref: ref("./pkg/apis/multicloudapps/v1.HelmReleaseRepo"),
+							Ref: ref("github.com/open-cluster-management/multicloud-operators-subscription-release/pkg/apis/apps/v1.HelmReleaseRepo"),
 						},
 					},
 					"spec": {
 						SchemaProps: spec.SchemaProps{
-							Ref: ref("./pkg/apis/multicloudapps/v1.HelmAppSpec"),
+							Ref: ref("github.com/open-cluster-management/multicloud-operators-subscription-release/pkg/apis/apps/v1.HelmAppSpec"),
 						},
 					},
 					"status": {
 						SchemaProps: spec.SchemaProps{
-							Ref: ref("./pkg/apis/multicloudapps/v1.HelmAppStatus"),
+							Ref: ref("github.com/open-cluster-management/multicloud-operators-subscription-release/pkg/apis/apps/v1.HelmAppStatus"),
 						},
 					},
 				},
 			},
 		},
 		Dependencies: []string{
-			"./pkg/apis/multicloudapps/v1.HelmAppSpec", "./pkg/apis/multicloudapps/v1.HelmAppStatus", "./pkg/apis/multicloudapps/v1.HelmReleaseRepo", "k8s.io/apimachinery/pkg/apis/meta/v1.ObjectMeta"},
+			"github.com/open-cluster-management/multicloud-operators-subscription-release/pkg/apis/apps/v1.HelmAppSpec", "github.com/open-cluster-management/multicloud-operators-subscription-release/pkg/apis/apps/v1.HelmAppStatus", "github.com/open-cluster-management/multicloud-operators-subscription-release/pkg/apis/apps/v1.HelmReleaseRepo", "k8s.io/apimachinery/pkg/apis/meta/v1.ObjectMeta"},
 	}
 }
 
-func schema_pkg_apis_app_v1_HelmReleaseRepo(ref common.ReferenceCallback) common.OpenAPIDefinition {
+func schema_pkg_apis_apps_v1_HelmReleaseRepo(ref common.ReferenceCallback) common.OpenAPIDefinition {
 	return common.OpenAPIDefinition{
 		Schema: spec.Schema{
 			SchemaProps: spec.SchemaProps{
@@ -89,7 +89,7 @@ func schema_pkg_apis_app_v1_HelmReleaseRepo(ref common.ReferenceCallback) common
 					"source": {
 						SchemaProps: spec.SchemaProps{
 							Description: "INSERT ADDITIONAL SPEC FIELDS - desired state of cluster Important: Run \"operator-sdk generate k8s\" to regenerate code after modifying this file Add custom validation using kubebuilder tags: https://book-v1.book.kubebuilder.io/beyond_basics/generating_crd.html Source holds the url toward the helm-chart",
-							Ref:         ref("./pkg/apis/multicloudapps/v1.Source"),
+							Ref:         ref("github.com/open-cluster-management/multicloud-operators-subscription-release/pkg/apis/apps/v1.Source"),
 						},
 					},
 					"chartName": {
@@ -122,6 +122,6 @@ func schema_pkg_apis_app_v1_HelmReleaseRepo(ref common.ReferenceCallback) common
 			},
 		},
 		Dependencies: []string{
-			"./pkg/apis/multicloudapps/v1.Source", "k8s.io/api/core/v1.ObjectReference"},
+			"github.com/open-cluster-management/multicloud-operators-subscription-release/pkg/apis/apps/v1.Source", "k8s.io/api/core/v1.ObjectReference"},
 	}
 }
