+++ pkg/controller/helmrelease/helmreleasemgr.go
@@ -23,7 +23,7 @@ import (
 	helmrelease "github.com/operator-framework/operator-sdk/pkg/helm/release"
 	"k8s.io/klog"
 
-	appv1 "github.com/open-cluster-management/multicloud-operators-subscription-release/pkg/apis/multicloudapps/v1"
+	appv1 "github.com/open-cluster-management/multicloud-operators-subscription-release/pkg/apis/apps/v1"
 	"github.com/open-cluster-management/multicloud-operators-subscription-release/pkg/utils"
 )
