+++ pkg/utils/helmrepoutils.go
@@ -46,7 +46,7 @@ import (
 	"k8s.io/helm/pkg/repo"
 	"k8s.io/klog"
 
-	appv1 "github.com/open-cluster-management/multicloud-operators-subscription-release/pkg/apis/multicloudapps/v1"
+	appv1 "github.com/open-cluster-management/multicloud-operators-subscription-release/pkg/apis/apps/v1"
 )
 
 //GetHelmRepoClient returns an *http.client to access the helm repo
